test(db-macros): split tests into two modules, improve real coverage
