+++ crates/db-macros/tests/proxy.rs
@@ -5,10 +5,11 @@ use std::sync::atomic::{AtomicU64, Ordering};
 
 use strata_db_macros::gen_proxy;
 use tokio::runtime::Handle;
-// `proc_macro2`/`quote`/`syn` are normal dependencies of `strata-db-macros` (the
-// proc-macro crate) and are visible to — but unused by — this integration target;
-// silence the workspace `unused_crate_dependencies` lint.
-use {proc_macro2 as _, quote as _, syn as _};
+// `proc_macro2`/`quote`/`syn` (normal deps of the proc-macro crate) and the
+// `tracing`/`tracing-subscriber` dev-deps (exercised by `tests/instrumentation.rs`) are
+// visible to — but unused by — this integration target; silence the workspace
+// `unused_crate_dependencies` lint.
+use {proc_macro2 as _, quote as _, syn as _, tracing as _, tracing_subscriber as _};
 
 /// Result alias mirroring the `DbResult<T>` shape used by real database traits.
 type DbResult<T> = Result<T, DbError>;
@@ -156,94 +157,4 @@ async fn clone_shares_underlying_state() {
 fn non_self_method_is_not_proxied() {
     // `describe` has no `&self` receiver, so only the trait method exists.
     assert_eq!(<MemCounter as CounterDb>::describe(), "counter");
-}
-
-// An instrumented trait: the `tracing_component` argument wraps each method's blocking
-// work in a tracing span. This exercises that the instrumented codegen compiles and
-// behaves identically to the uninstrumented variant.
-#[gen_proxy(error = DbError, tracing_component = "test:metric")]
-trait MetricDb: Send + Sync + 'static {
-    /// Records a value and returns the running total.
-    fn record(&self, value: u64) -> DbResult<u64>;
-
-    /// Returns the running total.
-    fn total(&self) -> DbResult<u64>;
-}
-
-impl MetricDb for MemCounter {
-    fn record(&self, value: u64) -> DbResult<u64> {
-        Ok(self.value.fetch_add(value, Ordering::SeqCst) + value)
-    }
-
-    fn total(&self) -> DbResult<u64> {
-        Ok(self.value.load(Ordering::SeqCst))
-    }
-}
-
-fn metric_proxy() -> MetricDbProxy {
-    let db = Arc::new(MemCounter {
-        value: AtomicU64::new(0),
-    });
-    MetricDbProxy::new(Handle::current(), db)
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn instrumented_variants_agree() {
-    let proxy = metric_proxy();
-
-    assert_eq!(proxy.record_blocking(5).unwrap(), 5);
-    assert_eq!(proxy.record_async(3).await.unwrap(), 8);
-    assert_eq!(proxy.total_chan().recv().await.unwrap(), 8);
-}
-
-// Captures `(span name, parent span name)` for every span created under the global
-// subscriber, used to assert cross-thread span parenting.
-static CAPTURED_SPANS: std::sync::Mutex<Vec<(String, Option<String>)>> =
-    std::sync::Mutex::new(Vec::new());
-
-struct CaptureLayer;
-
-impl<S> tracing_subscriber::Layer<S> for CaptureLayer
-where
-    S: tracing::Subscriber + for<'a> tracing_subscriber::registry::LookupSpan<'a>,
-{
-    fn on_new_span(
-        &self,
-        _attrs: &tracing::span::Attributes<'_>,
-        id: &tracing::span::Id,
-        ctx: tracing_subscriber::layer::Context<'_, S>,
-    ) {
-        if let Some(span) = ctx.span(id) {
-            let name = span.name().to_string();
-            let parent = span.parent().map(|p| p.name().to_string());
-            CAPTURED_SPANS.lock().unwrap().push((name, parent));
-        }
-    }
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-async fn instrumented_span_is_parented_to_caller() {
-    use tracing_subscriber::layer::SubscriberExt;
-
-    // Install a global subscriber so spans created on the `spawn_blocking` thread are
-    // recorded too. Only this test sets one; ignore the error if it is already set.
-    let subscriber = tracing_subscriber::registry().with(CaptureLayer);
-    let _ = tracing::subscriber::set_global_default(subscriber);
-
-    let proxy = metric_proxy();
-
-    // Issue the call from within a known parent span. `record_chan` captures the current
-    // span synchronously (here, `caller_parent`); the spawned shim re-enters it before
-    // creating the `record` span on the blocking thread.
-    let parent = tracing::info_span!("caller_parent");
-    let pending = parent.in_scope(|| proxy.record_chan(5));
-    assert_eq!(pending.recv().await.unwrap(), 5);
-
-    let captured = CAPTURED_SPANS.lock().unwrap();
-    assert!(
-        captured
-            .iter()
-            .any(|(name, parent)| name == "record" && parent.as_deref() == Some("caller_parent")),
-        "expected a `record` span parented to `caller_parent`, got: {captured:?}",
-    );
 }
